bdevperf: replace floats with doubles
Use double-precision floating point in the performance_dump() function,
since we can be dealing with large numbers (especially now that
io_completed is now uint64_t).

This isn't part of the I/O path, just used to print statistics, so
there should be no performance impact.  Also, the arguments passed to
printf() were already implicitly being promoted to double, so this
actually removes an extra conversion step.

+++ test/lib/bdev/bdevperf/bdevperf.c
@@ -548,8 +548,8 @@ performance_dump(int io_time_in_usec)
 {
 	uint32_t index;
 	unsigned lcore_id;
-	float io_per_second, mb_per_second;
-	float total_io_per_second, total_mb_per_second;
+	double io_per_second, mb_per_second;
+	double total_io_per_second, total_mb_per_second;
 	struct io_target *target;
 
 	total_io_per_second = 0;
@@ -561,7 +561,7 @@ performance_dump(int io_time_in_usec)
 			printf("\r Logical core: %u\n", lcore_id);
 		}
 		while (target != NULL) {
-			io_per_second = (float)target->io_completed * 1000000 /
+			io_per_second = (double)target->io_completed * 1000000 /
 					io_time_in_usec;
 			mb_per_second = io_per_second * g_io_size /
 					(1024 * 1024);
